Python/requirements.txt
mysql-connector-python
connexion
connexion[swagger-ui]
flask
flask-cors